[Passing] Implemented multiple items for auction.
Added end-to-end test for multiple item auctions.

Added support for multiple items in Main.

Fixed bug in MainWindow where the initial state of SniperTableModel was not set to constructor argument.

Added unit tests for SniperTableModelTest to test for multiple item insertions.

Changed API of ApplicationRunner to support multiple auctions.

Changed matcher in FakeAuctionServer to show a descriptive error message.